# Conflicts:
#	db_api.py
#	modules/start_dialogs.py
#	modules/users_classes.py
#	requirements.txt

+++ modules/timer.py
@@ -1,8 +1,7 @@
 import pytz
 import datetime as dt
 
-from telegram.ext import CallbackContext, ConversationHandler, \
-    DispatcherHandlerStop
+from telegram.ext import CallbackContext
 
 
 def remove_job_if_exists(name, context: CallbackContext):
@@ -23,9 +22,9 @@ def create_daily_notification(context: CallbackContext, **kwargs):
         # Добавляем задачу в очередь
 
         # Удаляем старую задачу с таким же именем
-        remove_job_if_exists(f'{chat_id} - {kwargs["name"]}', context)
+        remove_job_if_exists(f'{chat_id}-{kwargs["name"]}', context)
 
-        context.job_queue.run_daily(
+        job = context.job_queue.run_daily(
             callback=daily_task,
             # when=5,
             time=kwargs['time'],
@@ -39,6 +38,7 @@ def create_daily_notification(context: CallbackContext, **kwargs):
 
 
 def daily_task(context: CallbackContext):
+    # TODO запрос на проверку даты последней записи
     """Таски, которые выполняются ежедневно утром и вечером"""
     job = context.job
     data = job.context
@@ -55,34 +55,45 @@ def daily_task(context: CallbackContext):
     if user.msg_to_del:
         context.bot.delete_message(user.chat_id,
                                    user.msg_to_del.message_id)
-        remove_job_if_exists(user.rep_task_name, context)
+    remove_job_if_exists(user.rep_task_name, context)
 
     # Создание диалога для сбора данных
     user.notification_states[data['name']][
         user.state()[1]].pre_start(context, data)
 
-    user.rep_task_name = f'{user.chat_id}-{user.msg_to_del.message_id}'
+    user.rep_task_name = f'{user.chat_id}-rep_task'
 
     # Создаем новую циклическую задачу
     context.job_queue.run_repeating(
         callback=repeating_task,
-        interval=20,
-        last=dt.datetime.now(pytz.utc) + dt.timedelta(seconds=20*4),
-        # interval=data['task_data']['interval'],
-        # last=data['task_data']['last'],
+        # interval=20,
+        # last=dt.datetime.now(pytz.utc) + dt.timedelta(seconds=20*4),
+        interval=data['task_data']['interval'],
+        last=data['task_data']['last'],
         context=data,
         name=user.rep_task_name
     )
 
 
 def repeating_task(context: CallbackContext):
+    # TODO запрос на проверку даты последней записи
     job = context.job
     data = job.context
 
-    # Удаляем старое сообщение
-    context.bot.delete_message(data['user'].chat_id,
-                               data['user'].msg_to_del.message_id)
+    if data['user'].msg_to_del:
+        # Удаляем старое сообщение
+        context.bot.delete_message(data['user'].chat_id,
+                                   data['user'].msg_to_del.message_id)
 
     # Запускаем новое уведомление
     data['user'].notification_states[data['name']][
         data['user'].state()[1]].pre_start(context, data)
+
+
+def deleting_pre_start_msg_task(context):
+    job = context.job
+    data = job.context
+    try:
+        context.bot.delete_message(data['chat_id'], data['msg_id'])
+    except Exception as e:
+        pass
